requirements.txt
BeautifulSoup==3.2.1
filepath==0.1
html2text==2015.6.21
Jinja2==2.10
lxml==3.4.4
Markdown==2.6.11
MarkupSafe==1.0
PyYAML==3.11
requests==2.8.1
wheel==0.24.0
